+++ src/osgEarthUtil/LineOfSight.cpp
@@ -787,7 +787,7 @@ RadialLineOfSightNode::compute_line(osg::Node* node, bool backgroundThread)
     osgSim::LineOfSight los;
     los.setDatabaseCacheReadCallback(0);
 
-    for (int i = 0; i < _numSpokes; i++)
+    for (unsigned int i = 0; i < (unsigned int)_numSpokes; i++)
     {
         double angle = delta * (double)i;
         osg::Quat quat(angle, up );
@@ -798,7 +798,7 @@ RadialLineOfSightNode::compute_line(osg::Node* node, bool backgroundThread)
 
     los.computeIntersections(node);
 
-    for (int i = 0; i < _numSpokes; i++)
+    for (unsigned int i = 0; i < (unsigned int)_numSpokes; i++)
     {
         osg::Vec3d start = los.getStartPoint(i);
         osg::Vec3d end = los.getEndPoint(i);
@@ -931,7 +931,7 @@ RadialLineOfSightNode::compute_fill(osg::Node* node, bool backgroundThread)
     osgSim::LineOfSight los;
     los.setDatabaseCacheReadCallback(0);
 
-    for (int i = 0; i < _numSpokes; i++)
+    for (unsigned int i = 0; i < (unsigned int)_numSpokes; i++)
     {
         double angle = delta * (double)i;
         osg::Quat quat(angle, up );
@@ -942,7 +942,7 @@ RadialLineOfSightNode::compute_fill(osg::Node* node, bool backgroundThread)
 
     los.computeIntersections(node);
 
-    for (int i = 0; i < _numSpokes; i++)
+    for (unsigned int i = 0; i < (unsigned int)_numSpokes; i++)
     {
         //Get the current hit
         osg::Vec3d currEnd = los.getEndPoint(i);
